rename : user 위치
